Extract subscription logic
